--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,6 @@
 dist
 .env
 *.test
+
+# binary
+trufflehog