--- a/pkg/sources/sources.go
+++ b/pkg/sources/sources.go
@@ -26,8 +26,6 @@ type Chunk struct {
 	Data []byte
 	// Verify specifies whether any secrets in the Chunk should be verified.
 	Verify bool
-	// Keywords found within the chunk's Data.
-	Keywords map[string]bool
 }
 
 // Source defines the interface required to implement a source chunker.