### What this PR does

This PR adds a keyword optimization for string searching using the [Aho-Corasick algorithm](https://en.wikipedia.org/wiki/Aho%E2%80%93Corasick_algorithm).

Rough benchmarks:

```
# FileSystem Scanning
## Trufflehog repo:
Old: ./trufflehog filesystem . --no-verification --no-update  57.32s user 1.92s system 138% cpu 42.702 total
New: ./trufflehog filesystem . --no-verification --no-update  23.75s user 1.20s system 143% cpu 17.366 total

## Rails repo:
Old: ./trufflehog filesystem ../sandbox/benchmark-repos/rails --no-verification   51.07s user 3.08s system 379% cpu 14.263 total
New: ./trufflehog filesystem ../sandbox/benchmark-repos/rails --no-verification   14.12s user 2.38s system 184% cpu 8.939 total

## GitLab repo:
Old: ./trufflehog filesystem ../sandbox/benchmark-repos/gitlab --no-verification   746.48s user 23.35s system 802% cpu 1:35.89 total
Peak footprint GB (from MacOS activity monitor): 9.1
New: ./trufflehog filesystem ../sandbox/benchmark-repos/gitlab --no-verification   191.10s user 24.91s system 366% cpu 58.878 total
Peak footprint GB (from MacOS activity monitor): 8.0

## GitLab repo with Concurrency limited to 1:
Old: ./trufflehog filesystem ../sandbox/benchmark-repos/gitlab --no-verification    638.34s user 32.47s system 106% cpu 10:32.06 total
Peak footprint GB (from MacOS activity monitor): 6.8
New: ./trufflehog filesystem ../sandbox/benchmark-repos/gitlab --no-verification    175.53s user 32.21s system 117% cpu 2:56.88 total
Peak footprint GB (from MacOS activity monitor): 8.0

---

# Git scanning:
## Trufflehog repo:
Old: ./trufflehog git https://github.com/trufflesecurity/trufflehog.git    28.04s user 1.34s system 365% cpu 8.029 total
New: ./trufflehog git https://github.com/trufflesecurity/trufflehog.git    9.94s user 1.48s system 165% cpu 6.911 total

## Gitleaks repo:
Old: ./trufflehog git https://github.com/zricethezav/gitleaks.git --no-verificatio  24.70s user 0.72s system 420% cpu 6.050 total
New: ./trufflehog git https://github.com/zricethezav/gitleaks.git --no-verificatio  6.62s user 0.67s system 138% cpu 5.284 total

```

Quick Summary: Looks like memory impact is slightly better in some cases and slightly worse in others. I'd consider it a wash (note that I was doing some rough sampling via activity monitor and not using `pprof` which allows for more granular metrics). CPU/Timing on the other hand is improved. The most dramatic improvement came from limiting concurrency to 1 on the GitLab filesystem scan which resulted in a 3x speedup. In all cases we see a reduction in `real` time.